+++ test/unit/mongo_rep_admin/fetch_priority.py
@@ -110,9 +110,11 @@ class Server(object):
         self.auth = "auth"
         self.conf_file = "conffile"
         self.auth_db = "authentication_db"
-        self.use_arg = True
-        self.use_uri = False
         self.auth_mech = "SCRAM-SHA-1"
+        self.ssl_client_ca = None
+        self.ssl_client_cert = None
+        self.ssl_client_key = None
+        self.ssl_client_phrase = None
 
 
 class UnitTest(unittest.TestCase):